[Edit] fix errors

+++ app/controllers/public/post_comments_controller.rb
@@ -1,5 +1,4 @@
 class Public::PostCommentsController < ApplicationController
-  before_action :ensure_end_user, only: [:destroy]
 
   def create
     @post = Post.find(params[:post_id])
@@ -14,15 +13,14 @@ class Public::PostCommentsController < ApplicationController
   def destroy
     @post = Post.find(params[:post_id])
     post_comment = @post.post_comments.find(params[:id])
-    post_comment.destroy
+    if post_comment.end_user == current_end_user
+      post_comment.destroy
+    else
+      redirect_back(fallback_location: post_path(@post))
+    end
   end
 
   private
-
-  def ensure_end_user
-    @post.post_comment.end_user != current_end_user
-    redirect_to end_user_path(current_end_user)
-  end
 
 
   def post_comment_params
